+++ generator/internal/connection_generator.cc
@@ -82,19 +82,27 @@ Status ConnectionGenerator::GenerateHeader() {
                 // clang-format on
   );
 
-  // streaming updater functions
-  for (auto const& method : methods()) {
-    HeaderPrintMethod(
-        method,
-        {MethodPattern(
-            {// clang-format off
-   {"\n"
-    "void $service_name$$method_name$StreamingUpdater(\n"
-    "    $response_type$ const& response,\n"
-    "    $request_type$& request);\n"}
-     }, IsStreamingRead)},
-             // clang-format on
-        __FILE__, __LINE__);
+  // TODO(#8234): This is a special case for backwards compatibility of the
+  // streaming updated function.
+  if (vars().at("service_name") == "BigQueryRead") {
+    // streaming updater functions
+    for (auto const& method : methods()) {
+      HeaderPrintMethod(
+          method,
+          {MethodPattern(
+              {// clang-format off
+     {"\n"
+      "GOOGLE_CLOUD_CPP_DEPRECATED(\n"
+      "    \"applications should not need this.\"\n"
+      "    \" Please file a bug at https://github.com/googleapis/google-cloud-cpp\"\n"
+      "    \" if you do.\")"
+      "void $service_name$$method_name$StreamingUpdater(\n"
+      "    $response_type$ const& response,\n"
+      "    $request_type$& request);\n"}
+       }, IsStreamingRead)},
+               // clang-format on
+          __FILE__, __LINE__);
+    }
   }
 
   // Abstract interface Connection base class
@@ -232,6 +240,17 @@ Status ConnectionGenerator::GenerateCc() {
 
   auto result = CcOpenNamespaces();
   if (!result.ok()) return result;
+
+  if (vars().at("service_name") == "BigQueryRead") {
+    CcPrint(R"""(
+void BigQueryReadReadRowsStreamingUpdater(
+    google::cloud::bigquery::storage::v1::ReadRowsResponse const& response,
+    google::cloud::bigquery::storage::v1::ReadRowsRequest& request) {
+  return bigquery_internal::BigQueryReadReadRowsStreamingUpdater(response,
+                                                                 request);
+}
+)""");
+  }
 
   CcPrint(R"""(
 $connection_class_name$::~$connection_class_name$() = default;